+++ examples/count-reducer/src/lib.rs
