+++ examples/react/README.md
@@ -1,0 +1,3 @@
+# React + TypeScript + Vite + SQLSync
+
+This template provides a minimal setup to get SQLSync working with React and Vite.
