+++ examples/react/index.html
@@ -1,0 +1,15 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>SQLSync + Vite + React + TS</title>
+</head>
+
+<body>
+  <div id="root">if nothing renders, open the console</div>
+  <script type="module" src="/src/main.tsx"></script>
+</body>
+
+</html>
