--- lib/sqlsync-react/index.html
@@ -1,16 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>sqlsync-react tests index</title>
-</head>
-
-<body>
-    <ul>
-        <li><a href="/test/react-sanity.html">react-sanity.html</a></li>
-    </ul>
-</body>
-
-</html>
