+++ lib/sqlsync-react/package.json
@@ -11,38 +11,33 @@
   },
   "files": ["dist", "src"],
   "type": "module",
-  "main": "./dist/sqlsync-react.js",
-  "types": "./src/index.ts",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/sqlsync-react.js",
-      "require": "./dist/sqlsync-react.umd.cjs",
+      "import": "./src/index.ts",
+      "require": "./dist/index.js",
       "types": "./src/index.ts"
     }
   },
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build"
+    "build": "rollup --config"
   },
   "devDependencies": {
-    "@types/node": "^20.8.8",
-    "@types/react": "^18.2.31",
-    "@types/react-dom": "^18.2.14",
-    "@vitejs/plugin-react": "^4.1.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-typescript": "^11.1.5",
+    "rollup": "^3.29.4",
+    "tslib": "^2.6.2",
     "typescript": "^5.2.2",
-    "vite": "^4.5.0",
-    "vite-plugin-dts": "^3.6.1"
+    "@types/react": "^18.2.31",
+    "@types/react-dom": "^18.2.14"
   },
   "peerDependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "dependencies": {
-    "@orbitinghail/sqlsync-worker": "workspace:^",
-    "@scure/base": "^1.1.3",
     "fast-equals": "^5.0.1",
-    "fast-sha256": "^1.3.0"
+    "@orbitinghail/sqlsync-worker": "workspace:^"
   }
 }
