+++ lib/sqlsync-react/src/index.ts
@@ -1,10 +1,3 @@
-import { SQLSyncProvider } from "./context";
-import { createDocHooks, useConnectionStatus } from "./hooks";
-import { sql } from "./sql";
-import { DocType, Row } from "./sqlsync";
-import { serializeMutationAsJSON } from "./util";
+export { SQLSyncProvider } from "./context";
+export { createDocHooks, useConnectionStatus } from "./hooks";
 
-export { SQLSyncProvider, createDocHooks, serializeMutationAsJSON, sql, useConnectionStatus };
-export type { DocType, Row };
-
-// eof: this file only exports
