--- lib/sqlsync-react/test/react-sanity.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>sqlsync-react tests</title>
-</head>
-
-<body>
-    <div id="root">if nothing renders, open the console</div>
-    <script type="module" src="./react-sanity.tsx"></script>
-</body>
-
-</html>
