+++ lib/sqlsync-react/tsconfig.json
@@ -1,38 +1,23 @@
 {
   "compilerOptions": {
     "target": "esnext",
-    "useDefineForClassFields": true,
     "module": "esnext",
-    "lib": [
-      "ES6",
-      "DOM",
-      "DOM.Iterable",
-      "ES2021.WeakRef"
-    ],
-    "types": [
-      "vite/client"
-    ],
-    "skipLibCheck": true,
+    "moduleResolution": "Bundler",
+    "baseUrl": ".",
+    "lib": ["ES6", "DOM", "WebWorker", "ES2021.WeakRef"],
+
+    /* Compiling */
+    "allowJs": true,
+    "outDir": "dist",
+    "declaration": true,
+    "sourceMap": true,
     "jsx": "react-jsx",
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "allowImportingTsExtensions": true,
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true,
+
     /* Linting */
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": [
-    "src",
-    "test"
-  ],
-  "references": [
-    {
-      "path": "./tsconfig.node.json"
-    }
-  ]
+  "include": ["src", "node_modules/@orbitinghail/sqlsync-worker"]
 }
