+++ lib/sqlsync-worker/package.json
@@ -4,7 +4,14 @@
   "description": "SQLSync is a collaborative offline-first wrapper around SQLite. It is designed to synchronize web application state between users, devices, and the edge.",
   "homepage": "https://sqlsync.dev",
   "license": "Apache-2.0",
-  "keywords": ["sqlsync", "sql", "database", "sqlite", "offline-first", "local-first"],
+  "keywords": [
+    "sqlsync",
+    "sql",
+    "database",
+    "sqlite",
+    "offline-first",
+    "local-first"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/orbitinghail/sqlsync"
@@ -26,25 +33,27 @@
       "types": "./src/index.ts"
     },
     "./worker.js": "./dist/worker.js",
-    "./api.js": "./dist/api.js",
     "./worker.ts": "./src/worker.ts",
-    "./api.ts": "./src/api.ts",
     "./sqlsync.wasm": "./sqlsync-wasm/pkg/sqlsync_wasm_bg.wasm"
   },
   "scripts": {
+    "tsc": "tsc",
     "build-wasm": "cd sqlsync-wasm && wasm-pack build --target web --dev",
     "build-wasm-release": "cd sqlsync-wasm && wasm-pack build --target web --release",
     "build": "pnpm run build-wasm && rollup --config",
     "build-release": "pnpm run build-wasm-release && rollup --config"
   },
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^25.0.7",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-typescript": "^11.1.5",
+    "microbundle": "^0.15.1",
     "rollup": "^3.29.4",
     "tslib": "^2.6.2",
     "typescript": "^5.2.2"
   },
   "dependencies": {
-    "@scure/base": "^1.1.3"
+    "@scure/base": "^1.1.3",
+    "fast-sha256": "^1.3.0"
   }
 }
