+++ lib/sqlsync-worker/rollup.config.mjs
@@ -1,3 +1,4 @@
+import commonjs from "@rollup/plugin-commonjs";
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 import typescript from "@rollup/plugin-typescript";
 
@@ -9,10 +10,7 @@ const output = (name, entry) => ({
     sourcemap: true,
     name,
   },
-  plugins: [typescript(), nodeResolve()],
+  plugins: [commonjs(), typescript(), nodeResolve()],
 });
 
-export default [
-  output("SQLSyncWorkerTypes", "src/index.ts"),
-  output("SQLSyncWorker", "src/worker.ts"),
-];
+export default [output("SQLSync", "src/index.ts"), output("SQLSyncWorker", "src/worker.ts")];
