+++ pnpm-workspace.yaml
@@ -3,4 +3,4 @@ packages:
   - "demo/frontend"
   - "lib/sqlsync-react"
   - "lib/sqlsync-worker"
-  - "docs"
+  - "examples/react"
